feat: Move `getResourceFromImageId` to `FileUtil` class.
